test(core-state): move helper methods into utils folder and update test imports
