+++ __tests__/unit/core-state/__utils__/make-vote-transactions.ts
