+++ __tests__/unit/core-state/dpos/dpos-previous-round.test.ts
@@ -7,10 +7,11 @@ import { DposState } from "@packages/core-state/src/dpos/dpos";
 import { WalletRepository } from "@packages/core-state/src/wallets";
 import { IBlock } from "@packages/crypto/src/interfaces";
 
+import { buildDelegateAndVoteWallets } from "../__utils__/build-delegate-and-vote-balances";
+import { makeChainedBlocks } from "../__utils__/make-chained-block";
+import { makeVoteTransactions } from "../__utils__/make-vote-transactions";
 import { addTransactionsToBlock } from "../__utils__/transactions";
-import { makeChainedBlocks, makeVoteTransactions } from "../helper";
 import { setUp } from "../setup";
-import { buildDelegateAndVoteWallets } from "../__utils__/build-delegate-and-vote-balances"
 
 let dposState: DposState;
 let dposPreviousRoundStateProv: DposPreviousRoundStateProvider;
