--- __tests__/unit/core-state/helper/index.ts
@@ -1,2 +1,0 @@
-export * from "./make-chained-block";
-export * from "./make-vote-transactions";
