+++ __tests__/unit/core-state/state-builder.test.ts
@@ -3,8 +3,8 @@ import "jest-extended";
 import { Enums, Utils } from "@packages/core-kernel";
 import { StateBuilder } from "@packages/core-state/src/state-builder";
 import { WalletRepository } from "@packages/core-state/src/wallets";
-import { Managers } from "@packages/crypto";
 import { Sandbox } from "@packages/core-test-framework/src";
+import { Managers } from "@packages/crypto";
 
 import { setUp, setUpDefaults } from "./setup";
