+++ __tests__/unit/core-state/stores/transactions.test.ts
@@ -1,6 +1,6 @@
 import { TransactionStore } from "@packages/core-state/src/stores/transactions";
+import { Factories, FactoryBuilder } from "@packages/core-test-framework/src/factories";
 import { ITransaction } from "@packages/crypto/src/interfaces";
-import { FactoryBuilder, Factories } from "@packages/core-test-framework/src/factories";
 
 let factory: FactoryBuilder;
